+++ src/__tests__/depsGraph.test.ts
@@ -2,7 +2,7 @@
 
 import dedent from 'dedent';
 import * as babel from '@babel/core';
-import buildDepsGraph from '../babel/evaluate/depsGraph';
+import buildDepsGraph from '../babel/evaluate/graphBuilder';
 
 function _build(literal: TemplateStringsArray, ...placeholders: string[]) {
   const code = dedent(literal, ...placeholders);
