+++ src/babel/utils/peek.ts
@@ -1,0 +1,3 @@
+const peek = <T>(stack: T[], offset = 1): T => stack[stack.length - offset];
+
+export default peek;
